system-design-challenges-50/day16/requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
redis==3.5.3
celery==5.1.2
pydantic==1.8.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pytest==6.2.5
pytest-asyncio==0.15.1
alembic==1.7.3
httpx==0.19.0